Extensible mechanism to provide different serialization method
