+++ src/Shared/ISerializer.cs
@@ -1,0 +1,8 @@
+namespace Microsoft.Web.Redis
+{
+    public interface ISerializer
+    {
+        byte[] Serialize(object data);
+        object Deserialize(byte[] data);
+    }
+}
